Move plot method to base PixelAperture class
